alpha_agent_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain-openai==0.0.2
langgraph==0.0.20
langchain-core==0.1.10
supabase==2.0.3
yfinance==0.2.32
tushare==1.2.89
crawl4ai==0.3.0
playwright==1.40.0
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
